Act on the real clang-tidy findings, silence the rest with reasons
Ran clang-tidy against the native compile database rather than guessing from
inspection titles. Seven findings were worth acting on:

- gammaCorrect rounded by adding 0.5 and truncating, which is wrong at the
  representable edges. Now lroundf.
- Param::set clamped with two ifs; std::clamp says it in one line, and the
  store is now skipped when the value did not change.
- Frame's fill/fade/scale/clear are const: a Frame is a view, and they change
  the pixels it refers to, not the Frame — the rule std::span follows. at(),
  operator[] and raw() follow, plus [[nodiscard]].
- Fire indexed heat_ through an int multiplication widened to ptrdiff_t after
  the fact; widened once at the source instead.
- Registry::add and Registrar are noexcept, so static registration is provably
  non-throwing.
- Parenthesised the index arithmetic in Matrix, Frame and Fire.
- Named the unused parameter in Effect::begin; dropped an unused Arduino.h.

Three were declined and the reason recorded in .clang-tidy, because they will
otherwise be re-litigated every time someone opens the project:

- Functional-style casts stay. On a fundamental type uint8_t(x) is exactly
  static_cast and cannot decay into reinterpret_cast, which is the hazard the
  rule guards. The project rule is now written down instead.
- Param's set() stays non-const even though std::atomic::store is const —
  const would be true for the compiler and false for the reader.
- readability-avoid-const-params-in-decls and CLion's "parameter can be made
  const" want opposite things; the project sides with the shorter declaration.

clang-tidy now reports zero warnings on core and effects. Both targets build,
host tests pass.

+++ src/effects/Fire.cpp
@@ -44,7 +44,7 @@ private:
         const uint8_t h = f.height();
 
         for (uint8_t x = 0; x < w; ++x) {
-            uint8_t* col = &heat_[uint16_t(x) * h];
+            uint8_t* col = &heat_[size_t(x) * h];
 
             // Cool every cell a little; the taller the matrix, the less each
             // step may take, or the flame never reaches the top.
@@ -67,7 +67,7 @@ private:
         const uint8_t base = uint8_t(hue);
         for (uint8_t x = 0; x < f.width(); ++x)
             for (uint8_t y = 0; y < f.height(); ++y) {
-                const uint8_t t = heat_[uint16_t(x) * f.height() + y];
+                const uint8_t t = heat_[(size_t(x) * f.height()) + y];
                 // HeatColor gives the classic black-red-yellow-white ramp;
                 // the hue parameter tints it towards green or violet flame.
                 CRGB c = HeatColor(t);
